user-service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = "Control Agent User Service"
authors = ["Barath Kandappan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.11"
sqlalchemy = "^2.0.39"
python-consul = "^1.1.0"
uvicorn = "^0.34.0"
dotenv = "^0.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
